future proof return size

+++ features/FEATURE_BLE/ble/gap/Gap.h
@@ -548,7 +548,7 @@ public:
      *
      * @return Maximum advertising data length you may set if advertising set is active.
      */
-    virtual uint8_t getMaxActiveSetAdvertisingDataLength();
+    virtual uint16_t getMaxActiveSetAdvertisingDataLength();
 
     /** Create an advertising set and apply the passed in parameters. The handle returned
      *  by this function must be used for all other calls that accept an advertising handle.
