+++ features/FEATURE_BLE/source/generic/GenericGap.cpp
@@ -2015,7 +2015,7 @@ uint16_t GenericGap::getMaxConnectableAdvertisingDataLength()
     return _pal_gap.get_maximum_connectable_advertising_data_length();
 }
 
-uint8_t GenericGap::getMaxActiveSetAdvertisingDataLength()
+uint16_t GenericGap::getMaxActiveSetAdvertisingDataLength()
 {
     useVersionTwoAPI();
     return _pal_gap.get_maximum_hci_advertising_data_length();
